# server/stycoon/model_attack_command.go
/*
 * Space Tycoon
 *
 * Space Tycoon server.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package stycoon

type AttackCommand struct {

	Type string `json:"type"`

	Target *string `json:"target"`
}

// AssertAttackCommandRequired checks if the required fields are not zero-ed
func AssertAttackCommandRequired(obj AttackCommand) error {
	elements := map[string]interface{}{
		"type": obj.Type,
		"target": obj.Target,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertRecurseAttackCommandRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of AttackCommand (e.g. [][]AttackCommand), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseAttackCommandRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aAttackCommand, ok := obj.(AttackCommand)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertAttackCommandRequired(aAttackCommand)
	})
}