# server/stycoon/model_rename_command.go
/*
 * Space Tycoon
 *
 * Space Tycoon server.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package stycoon

type RenameCommand struct {

	Type string `json:"type"`

	Name *string `json:"name"`
}

// AssertRenameCommandRequired checks if the required fields are not zero-ed
func AssertRenameCommandRequired(obj RenameCommand) error {
	elements := map[string]interface{}{
		"type": obj.Type,
		"name": obj.Name,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertRecurseRenameCommandRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of RenameCommand (e.g. [][]RenameCommand), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseRenameCommandRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aRenameCommand, ok := obj.(RenameCommand)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertRenameCommandRequired(aRenameCommand)
	})
}